Move mixins from models/base to models/mixins
