+++ etna/models/catboost.py
@@ -8,10 +8,10 @@ from catboost import Pool
 from deprecated import deprecated
 
 from etna.models.base import BaseAdapter
-from etna.models.base import MultiSegmentModelMixin
 from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
-from etna.models.base import NonPredictionIntervalContextIgnorantModelMixin
-from etna.models.base import PerSegmentModelMixin
+from etna.models.mixins import MultiSegmentModelMixin
+from etna.models.mixins import NonPredictionIntervalContextIgnorantModelMixin
+from etna.models.mixins import PerSegmentModelMixin
 
 
 class _CatBoostAdapter(BaseAdapter):
