+++ etna/models/sklearn.py
@@ -6,10 +6,10 @@ import pandas as pd
 from sklearn.base import RegressorMixin
 
 from etna.models.base import BaseAdapter
-from etna.models.base import MultiSegmentModelMixin
 from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
-from etna.models.base import NonPredictionIntervalContextIgnorantModelMixin
-from etna.models.base import PerSegmentModelMixin
+from etna.models.mixins import MultiSegmentModelMixin
+from etna.models.mixins import NonPredictionIntervalContextIgnorantModelMixin
+from etna.models.mixins import PerSegmentModelMixin
 
 
 class _SklearnAdapter(BaseAdapter):
