+++ tests/test_models/test_mixins.py
@@ -1,0 +1,50 @@
+from unittest.mock import MagicMock
+
+import pytest
+
+from etna.models.mixins import MultiSegmentModelMixin
+from etna.models.mixins import PerSegmentModelMixin
+
+
+@pytest.fixture()
+def regression_base_model_mock():
+    cls = MagicMock()
+    del cls.forecast
+
+    model = MagicMock()
+    model.__class__ = cls
+    del model.forecast
+    return model
+
+
+@pytest.fixture()
+def autoregression_base_model_mock():
+    cls = MagicMock()
+
+    model = MagicMock()
+    model.__class__ = cls
+    return model
+
+
+@pytest.mark.parametrize("mixin_constructor", [PerSegmentModelMixin, MultiSegmentModelMixin])
+@pytest.mark.parametrize(
+    "base_model_name, called_method_name, expected_method_name",
+    [
+        ("regression_base_model_mock", "_forecast", "predict"),
+        ("autoregression_base_model_mock", "_forecast", "forecast"),
+        ("regression_base_model_mock", "_predict", "predict"),
+        ("autoregression_base_model_mock", "_predict", "predict"),
+    ],
+)
+def test_calling_private_prediction(
+    base_model_name, called_method_name, expected_method_name, mixin_constructor, request
+):
+    base_model = request.getfixturevalue(base_model_name)
+    ts = MagicMock()
+    mixin = mixin_constructor(base_model=base_model)
+    mixin._make_predictions = MagicMock()
+    to_call = getattr(mixin, called_method_name)
+    to_call(ts=ts)
+    mixin._make_predictions.assert_called_once_with(
+        ts=ts, prediction_method=getattr(base_model.__class__, expected_method_name)
+    )
